Move elf_read_soname to test file
