+++ tests/unit/test_elfutils.py
@@ -12,7 +12,6 @@ from auditwheel.elfutils import (
     elf_find_versioned_symbols,
     elf_read_dt_needed,
     elf_read_rpaths,
-    elf_read_soname,
     elf_references_pyfpe_jbuf,
     get_undefined_symbols,
 )
@@ -28,50 +27,6 @@ class MockSymbol(dict[str, Any]):
     @property
     def name(self):
         return self._name
-
-
-@patch("auditwheel.elfutils.ELFFile")
-class TestElfReadSoname:
-    def test_missing_section(self, elffile_mock, tmp_path):
-        fake = tmp_path / "fake.so"
-        fake.touch()
-
-        # GIVEN
-        elffile_mock.return_value.get_section_by_name.return_value = None
-
-        # THEN
-        assert elf_read_soname(fake) is None
-
-    def test_missing_tag(self, elffile_mock, tmp_path):
-        fake = tmp_path / "fake.so"
-        fake.touch()
-
-        # GIVEN
-        section_mock = Mock()
-        section_mock.iter_tags.return_value = []
-        elffile_mock.return_value.get_section_by_name.return_value = section_mock
-
-        # THEN
-        assert elf_read_soname(fake) is None
-
-    def test_read_soname(self, elffile_mock, tmp_path):
-        fake = tmp_path / "fake.so"
-        fake.touch()
-
-        # GIVEN
-        tag1 = Mock(needed="libz.so")  # Non-SONAME tags should be ignored
-        tag1.entry.d_tag = "DT_NEEDED"
-        tag2 = Mock(soname="libfoo.so")
-        tag2.entry.d_tag = "DT_SONAME"
-        tag3 = Mock(soname="libbar.so")  # Subsequent SONAME tags should be ignored
-        tag3.entry.d_tag = "DT_SONAME"
-
-        section_mock = Mock()
-        section_mock.iter_tags.return_value = [tag1, tag2, tag3]
-        elffile_mock.return_value.get_section_by_name.return_value = section_mock
-
-        # THEN
-        assert elf_read_soname(fake) == "libfoo.so"
 
 
 @patch("auditwheel.elfutils.ELFFile")
